requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
cffi==1.16.0
colorama==0.4.6
cryptography==42.0.2
Django==4.2.9
django-qr-code==4.0.1
mysqlclient==2.2.4
numpy==1.26.4
param==2.0.2
pillow==10.2.0
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pypng==0.20220715.0
qrcode==7.4.2
segno==1.6.1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1